# geocell/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GeoCell API
  version: 1.0.0
  description: Endpoints de autenticação e administração para GeoCell.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/departments:
    get:
      tags: [Departments]
      summary: Listar departamentos
      description: Requer utilizador autenticado com perfil superuser.
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: Lista de departamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      tags: [Departments]
      summary: Criar departamento
      description: Requer utilizador autenticado com perfil superuser.
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDepartmentRequest'
      responses:
        '201':
          description: Departamento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/departments/{id}:
    put:
      tags: [Departments]
      summary: Atualizar departamento
      description: Requer utilizador autenticado com perfil superuser.
      security:
        - SessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDepartmentRequest'
      responses:
        '200':
          description: Departamento atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

    delete:
      tags: [Departments]
      summary: Remover departamento
      description: Requer utilizador autenticado com perfil superuser. Falha com conflito se existirem users atribuídos ou operations ligadas ao departamento.
      security:
        - SessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Departamento removido
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/users:
    get:
      tags: [Users]
      summary: Listar utilizadores
      description: Requer utilizador autenticado com perfil superuser.
      security:
        - SessionCookieAuth: []
      responses:
        '200':
          description: Lista de utilizadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      tags: [Users]
      summary: Criar utilizador
      description: Requer utilizador autenticado com perfil superuser.
      security:
        - SessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Utilizador criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Obter utilizador por ID
      description: Requer utilizador autenticado com perfil superuser.
      security:
        - SessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Utilizador encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      tags: [Users]
      summary: Atualizar utilizador por ID
      description: Requer utilizador autenticado com perfil superuser. O campo `password` é opcional; quando omitido, a password atual mantém-se.
      security:
        - SessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Utilizador atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
    delete:
      tags: [Users]
      summary: Remover utilizador por ID
      description: Requer utilizador autenticado com perfil superuser.
      security:
        - SessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Utilizador removido
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: session_key
  responses:
    ValidationError:
      description: Pedido inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Sem permissões para a operação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflito de dados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateDepartmentRequest:
      type: object
      required:
        - name
        - haveOperations
      properties:
        name:
          type: string
          maxLength: 150
        haveOperations:
          type: boolean
    UpdateDepartmentRequest:
      type: object
      required:
        - name
        - haveOperations
      properties:
        name:
          type: string
          maxLength: 150
        haveOperations:
          type: boolean
    DepartmentResponse:
      type: object
      required:
        - id
        - name
        - haveOperations
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        haveOperations:
          type: boolean
    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - name
        - email
        - departmentId
      properties:
        username:
          type: string
          maxLength: 150
        password:
          type: string
          maxLength: 128
        name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        departmentId:
          type: integer
          format: int64
        isSuperuser:
          type: boolean
          default: false
        isAnalyst:
          type: boolean
          default: false
        isOperationAdmin:
          type: boolean
          default: false
        mapType:
          type: string
          enum: [standard, satellite, hybrid, terrain]
          default: standard
        showGrid:
          type: boolean
          default: false
        showCounties:
          type: boolean
          default: false
    UpdateUserRequest:
      type: object
      required:
        - username
        - name
        - email
        - departmentId
      properties:
        username:
          type: string
          maxLength: 150
        password:
          type: string
          maxLength: 128
          nullable: true
        name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        departmentId:
          type: integer
          format: int64
        isSuperuser:
          type: boolean
          default: false
        isAnalyst:
          type: boolean
          default: false
        isOperationAdmin:
          type: boolean
          default: false
        mapType:
          type: string
          enum: [standard, satellite, hybrid, terrain]
          default: standard
        showGrid:
          type: boolean
          default: false
        showCounties:
          type: boolean
          default: false
    UserResponse:
      type: object
      required:
        - id
        - username
        - name
        - email
        - isSuperuser
        - isAnalyst
        - isOperationAdmin
        - mapType
        - showGrid
        - showCounties
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        departmentId:
          type: integer
          format: int64
          nullable: true
        isSuperuser:
          type: boolean
        isAnalyst:
          type: boolean
        isOperationAdmin:
          type: boolean
        mapType:
          type: string
          enum: [standard, satellite, hybrid, terrain]
        showGrid:
          type: boolean
        showCounties:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string